Split browser automation panels
